+++ src/main/java/com/jandi/plan_backend/commu/controller/CommentController.java
@@ -1,7 +1,10 @@
 package com.jandi.plan_backend.commu.controller;
 
-import com.jandi.plan_backend.commu.dto.*;
-import com.jandi.plan_backend.commu.service.CommunityService;
+import com.jandi.plan_backend.commu.dto.CommentReqDTO;
+import com.jandi.plan_backend.commu.dto.CommentRespDTO;
+import com.jandi.plan_backend.commu.dto.ParentCommentDTO;
+import com.jandi.plan_backend.commu.dto.repliesDTO;
+import com.jandi.plan_backend.commu.service.CommentService;
 import com.jandi.plan_backend.security.JwtTokenProvider;
 import org.springframework.data.domain.Page;
 import org.springframework.http.ResponseEntity;
@@ -11,41 +14,13 @@ import java.util.Map;
 
 @RestController
 @RequestMapping("/api/community")
-public class CommunityController {
-
-    private final CommunityService communityService;
+public class CommentController {
+    private final CommentService commentService;
     private final JwtTokenProvider jwtTokenProvider;
 
-    public CommunityController(CommunityService communityService, JwtTokenProvider jwtTokenProvider) {
-        this.communityService = communityService;
+    public CommentController(CommentService commentService, JwtTokenProvider jwtTokenProvider) {
+        this.commentService = commentService;
         this.jwtTokenProvider = jwtTokenProvider;
-    }
-
-    /** 게시물 목록 조회 API*/
-    @GetMapping("/posts")
-    public Map<String, Object> getPosts(
-            @RequestParam(defaultValue = "0") int page,
-            @RequestParam(defaultValue = "10") int size
-    ){
-        Page<CommunityListDTO> postsPage = communityService.getAllPosts(page, size);
-
-        return Map.of(
-                "pageInfo", Map.of(
-                        "currentPage", postsPage.getNumber(),  // 현재 페이지 번호
-                        "currentSize", postsPage.getContent().size(), //현재 페이지 리스트 갯수
-                        "totalPages", postsPage.getTotalPages(),  // 전체 페이지 번호 개수
-                        "totalSize", postsPage.getTotalElements() // 전체 게시물 리스트 개수
-                ),
-                "items", postsPage.getContent()   // 현재 페이지의 게시물 데이터
-        );
-    }
-
-    /** 특정 게시물 조회 API*/
-    @GetMapping("/posts/{postId}")
-    public Map<String, Object> getPosts(
-            @PathVariable Integer postId //경로 변수로 게시물 고유 번호 받기
-    ){
-        return Map.of("items", communityService.getSpecPost(postId));
     }
 
     /** 댓글 조회 API */
@@ -55,7 +30,7 @@ public class CommunityController {
             @RequestParam(defaultValue = "0") int page,
             @RequestParam(defaultValue = "10") int size
     ){
-        Page<ParentCommentDTO> parentCommentsPage = communityService.getAllComments(postId, page, size);
+        Page<ParentCommentDTO> parentCommentsPage = commentService.getAllComments(postId, page, size);
 
         return Map.of(
                 "pageInfo", Map.of(
@@ -75,7 +50,7 @@ public class CommunityController {
             @RequestParam(defaultValue = "0") int page,
             @RequestParam(defaultValue = "10") int size
     ){
-        Page<repliesDTO> repliesPage = communityService.getAllReplies(commentId, page, size);
+        Page<repliesDTO> repliesPage = commentService.getAllReplies(commentId, page, size);
 
         return Map.of(
                 "pageInfo", Map.of(
@@ -86,21 +61,6 @@ public class CommunityController {
                 ),
                 "items", repliesPage.getContent()
         );
-    }
-
-    /** 게시글 작성 API */
-    @PostMapping("/posts")
-    public ResponseEntity<?> writePost(
-            @RequestHeader("Authorization") String token, // 헤더의 Authorization에서 JWT 토큰 받기
-            @RequestBody CommunityReqDTO postDTO // JSON 형식으로 게시글 작성 정보 받기
-    ){
-        // Jwt 토큰으로부터 유저 이메일 추출
-        String jwtToken = token.replace("Bearer ", "");
-        String userEmail = jwtTokenProvider.getEmail(jwtToken);
-
-        // 게시글 저장 및 반환
-        CommunityRespDTO savedPost = communityService.writePost(postDTO, userEmail);
-        return ResponseEntity.ok(savedPost);
     }
 
     /** 댓글 작성 API */
@@ -116,7 +76,7 @@ public class CommunityController {
         String userEmail = jwtTokenProvider.getEmail(jwtToken);
 
         // 댓글 저장 및 반환
-        CommentRespDTO savedComment = communityService.writeComment(commentDTO, postId, userEmail);
+        CommentRespDTO savedComment = commentService.writeComment(commentDTO, postId, userEmail);
         return ResponseEntity.ok(savedComment);
     }
 
@@ -133,24 +93,8 @@ public class CommunityController {
         String userEmail = jwtTokenProvider.getEmail(jwtToken);
 
         // 댓글 저장 및 반환
-        CommentRespDTO savedComment = communityService.writeReplies(commentDTO, commentId, userEmail);
+        CommentRespDTO savedComment = commentService.writeReplies(commentDTO, commentId, userEmail);
         return ResponseEntity.ok(savedComment);
-    }
-
-    /** 게시물 수정 API */
-    @PatchMapping("/posts/{postId}")
-    public ResponseEntity<?> updatePost(
-            @PathVariable Integer postId, //경로 변수로 변경할 게시글 아이디 받기
-            @RequestHeader("Authorization") String token, // 헤더의 Authorization에서 JWT 토큰 받기
-            @RequestBody CommunityReqDTO postDTO // JSON 형식으로 게시글 작성 정보 받기
-    ){
-        // Jwt 토큰으로부터 유저 이메일 추출
-        String jwtToken = token.replace("Bearer ", "");
-        String userEmail = jwtTokenProvider.getEmail(jwtToken);
-
-        // 게시물 수정 및 반환
-        CommunityRespDTO updatedPost = communityService.updatePost(postDTO, postId, userEmail);
-        return ResponseEntity.ok(updatedPost);
     }
 
     /** 댓글 및 답글 수정 API */
@@ -165,7 +109,7 @@ public class CommunityController {
         String userEmail = jwtTokenProvider.getEmail(jwtToken);
 
         // 게시물 수정 및 반환
-        CommentRespDTO updatedPost = communityService.updateComment(commentDTO, commentId, userEmail);
+        CommentRespDTO updatedPost = commentService.updateComment(commentDTO, commentId, userEmail);
         return ResponseEntity.ok(updatedPost);
     }
 }
